Add support for radio firmware recovery
Extend original recovery implementation with sysfs parameter indicating need for recovery.
Parameter: restart_required
   Value:  0   Normal operation
   Value:  1   Restart required

Host platform can monitor restart_required and perform whatever operations are necessary to reset module and reload driver.
This implementation is intended to be used on platforms that are unable to leverage the builtin recovery mechanism.

Bug: 11298