USB: Command Corruption
Modified usb RX command handler to validate that command response is the response to the currently pending command prior to copying to the shared cmd buffer and returning from the wait.
This fixes issues that occur when multiple command responses are unexpectedly received from the firmware.
Note that PCI and SDIO already check for matching command/response, USB was the outlier.

Changed command timeout from 30 seconds to 12 seconds, in line with SDIO and PCI timeouts.
Minor cleanup in wait_completed routine for readability.

Bug: 16802