Resolved kernel panic after driver unloaded
If radio is configured for world mode and using adaptive_ww, it is possible for a kernel panic
to occur if radio is unloaded and host is left in this state for a long period of time (e.g. 60 minutes).

Problem occurs because of a race condtion between AWM and disabling of timers.

Bug: 17168